+++ contracts/core/interfaces/IPoolRouter.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.15;
 interface IPoolRouter {
     function isReservedMode() external view returns (bool);
 
-    function getReservedMultiplier() external view returns (uint);
+    function getReservedMultiplier() external view returns (uint256);
 
-    function poolUtilisation(address _token, uint16 _index) external view returns (uint);
+    function poolUtilisation(address _token, uint16 _index) external view returns (uint256);
 }
